Fix bugs with conversion between fixed point

+++ core/src/hw/gpu/engine3d/math.rs
@@ -201,7 +201,7 @@ impl FixedPoint {
     pub fn zero() -> Self { FixedPoint(0) }
     pub fn from_mul(val: i64) -> Self { FixedPoint((val >> 12) as i32) }
     pub fn from_num(val: i32) -> Self { FixedPoint(val << 12) }
-    pub fn from_frac6(val: u16) -> Self { FixedPoint((val as i32) << 6) }
+    pub fn from_frac6(val: u16) -> Self { FixedPoint((val << 6) as i16 as i32) }
     pub fn from_frac9(val: u16) -> Self { FixedPoint((((val << 6) as i16) >> 3) as i32) }
     pub fn from_frac12(val: i32) -> Self { FixedPoint(val) }
     pub fn num(&self) -> usize { (self.0 >> 12) as usize }
